requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-filter==21.1
django-simple-history==3.0.0
django-cors-headers==3.10.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.2
python-decouple==3.5
django-import-export==2.6.1
django-material-admin==1.8.5
django-admin-honeypot==1.1.0
django-simple-history==3.0.0
